pyproject.toml
[project]
name = "ftm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "docling>=2.15.1",
    "fastapi>=0.115.7",
    "gradio>=5.13.1",
    "llama-cpp-python>=0.3.6",
    "outlines>=0.1.13",
    "sentencepiece>=0.2.0",
    "sqlmodel>=0.0.22",
    "torch>=2.5.1",
    "transformers>=4.48.1",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "ruff>=0.9.3",
]
